queryAndMergeS3, update findS3Keys and duration type

+++ app/src/main/java/com/example/test/App.java
@@ -6,11 +6,11 @@ public class App {
     public  Drawable appIcon;
     public  String appName;
     public  int usagePercentage;
-    public int durationInt;
+    public long durationInt;
     public  String usageDuration;
 
 
-    public App(Drawable appIcon, String appName, int usagePercentage, int durationInt, String usageDuration) {
+    public App(Drawable appIcon, String appName, int usagePercentage, long durationInt, String usageDuration) {
         this.appIcon = appIcon;
         this.appName = appName;
         this.usagePercentage = usagePercentage;
